All API files
As of now I have 22 specs passing, I don't have anymore time to work on this. Uploading the repo with the API shortly. If possible can you guys fix those issues I left the Starting part of delete in there as it's commented out on line 59-60.
ALSO
if it helps when you run the server it says getRoutineActivityByID is not a function. pretty sure I missed a really simply line somewhere. but uploading now